requirements.txt
asgiref==3.8.1
Django==4.2.16
psycopg2-binary==2.9.10
sqlparse==0.5.1
typing_extensions==4.12.2
whitenoise==6.8.2
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
python-dotenv